openmp now but still super slow

+++ src/test/csv.h
@@ -1,2 +1,2 @@
 #include<stdio.h>
-void write_to_csv(float *arr, int arr_size, char *filename);
+void write_to_csv(double *arr, int arr_size, char *filename);
